TG-XXX fix(kycBanking): remove banking added.

+++ ofi-kyc/my-requests/request/steps/identification/company-information.component.ts
@@ -243,10 +243,12 @@ export class CompanyInformationComponent implements OnInit, OnDestroy {
     removeBeneficiary(i) {
         let control = this.form.get('beneficiaries') as FormArray;
 
-        this.identificationService.deleteBeneficiary(
-            control.at(i).value.kycID,
-            control.at(i).value.companyBeneficiariesID
-        );
+        if (control.at(i).value.companyBeneficiariesID !== '') {
+            this.identificationService.deleteBeneficiary(
+                control.at(i).value.kycID,
+                control.at(i).value.companyBeneficiariesID
+            );
+        }
 
         control.removeAt(i);
